נוספה האפשרות להוסיף את כל פרטי העובד בד נפרד. וכן הצגת פרטי העובד בד נפרד. הכל עובד!

+++ js/employeesScript.js
@@ -4,7 +4,7 @@ function CreateTble(){
     let str="";
     for(let line of raw_data){
         str+="<tr>";
-        str+=`<td><button class="btn" onclick="editLine(${line.EmployeeID})">עדכן</button></td>`;
+        str+=`<td><button class="btn" onclick="editLine(${line.EmployeeID})">הצג</button></td>`;
         str+=`<td>${line.EmployeeID}</td>`;
         str+=`<td>${line.FullName}</td>`;
         str += `<td><img src="${line.ImageURL}" alt="${line.FullName}" width="30px" height="30px"></td>`;
@@ -20,35 +20,22 @@ async function getList() {
     console.log(raw_data);
     CreateTble();
 }
-
-
 async function deleteLine(id) {
+    var confirmation = window.confirm("האם אתה בטוח שברצונך להמשיך?");
+    if (confirmation){
     let response = await fetch(`/emp/Delete/${id}`,{
             method: 'DELETE',
         }
-    );
-    getList();
-}
-async function editLine(id) {
-    let objToServer={};
-    objToServer.id=id;
-    objToServer.firstName = document.getElementById("firstName").value;
-    objToServer.lastName = document.getElementById("lastName").value;
-    let response = await fetch('/emp/Update', {
-            method: 'PATCH',
-            headers: {
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify(objToServer)
-        }
-    );
+    );}
     getList();
 }
 function updateTime() {
     document.getElementById("title").innerHTML="רשימת העובדים";
-    document.querySelector(".signature-image").src="photos/my_employees.png";
+    document.getElementById("addEmp").src="/photos/add_emp.jpg";
+    document.getElementById("timeClock").src="/photos/my_clock.png";
 }
-
-
+async function editLine(id){
+    window.location.href = `http://localhost:5656/emp/EmpCard?id=${id}`;
+}
 getList();
 updateTime();
